rename directory for the current implementation
goal is to move everything in here which comprises the current "backend" part of the implementation
of autocomplete
