+++ src/services/ghost/GhostProvider.ts
@@ -2,14 +2,14 @@ import crypto from "crypto"
 import * as vscode from "vscode"
 import { t } from "../../i18n"
 import { GhostDocumentStore } from "./GhostDocumentStore"
-import { GhostStreamingParser } from "./GhostStreamingParser"
-import { AutoTriggerStrategy } from "./strategies/AutoTriggerStrategy"
+import { GhostStreamingParser } from "./classic-auto-complete/GhostStreamingParser"
+import { AutoTriggerStrategy } from "./classic-auto-complete/AutoTriggerStrategy"
 import { GhostModel } from "./GhostModel"
 import { GhostSuggestionContext, contextToAutocompleteInput, extractPrefixSuffix } from "./types"
 import { GhostStatusBar } from "./GhostStatusBar"
-import { GhostSuggestionsState } from "./GhostSuggestions"
+import { GhostSuggestionsState } from "./classic-auto-complete/GhostSuggestions"
 import { GhostCodeActionProvider } from "./GhostCodeActionProvider"
-import { GhostInlineCompletionProvider } from "./GhostInlineCompletionProvider"
+import { GhostInlineCompletionProvider } from "./classic-auto-complete/GhostInlineCompletionProvider"
 import { GhostServiceSettings, TelemetryEventName } from "@roo-code/types"
 import { ContextProxy } from "../../core/config/ContextProxy"
 import { ProviderSettingsManager } from "../../core/config/ProviderSettingsManager"
