+++ src/services/ghost/classic-auto-complete/AutoTriggerStrategy.ts
@@ -1,5 +1,5 @@
 import { AutocompleteInput } from "../types"
-import { CURSOR_MARKER } from "../ghostConstants"
+import { CURSOR_MARKER } from "./ghostConstants"
 import { isCommentLine, cleanComment } from "./CommentHelpers"
 import type { TextDocument, Range } from "vscode"
