+++ src/services/ghost/classic-auto-complete/CommentHelpers.ts
