+++ src/services/ghost/classic-auto-complete/GhostSuggestions.ts
