+++ src/services/ghost/classic-auto-complete/__tests__/AutoTriggerStrategy.test.ts
