+++ src/services/ghost/classic-auto-complete/ghostConstants.ts
